test: 기존 Fake 객체, 실제 구현으로 대체
