+++ src/main/resources/application.yml
@@ -1,7 +1,24 @@
 spring:
+  config:
+    activate:
+      on-profile: "default"
   h2:
     console:
       enabled: true
       path: /h2-console
   datasource:
     url: jdbc:h2:mem:database
+
+---
+spring:
+  config:
+    activate:
+      on-profile: "test"
+  h2:
+    console:
+      enabled: true
+      path: /h2-console-test
+  datasource:
+    url: jdbc:h2:mem:database_test
+server:
+  port: 8081
