+++ src/test/java/roomescape/controller/ReservationTimeControllerTest.java
@@ -1,25 +1,31 @@
 package roomescape.controller;
 
 import static org.hamcrest.Matchers.is;
-import static org.junit.jupiter.api.Assertions.*;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import java.time.LocalDate;
 import java.util.HashMap;
 import java.util.Map;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.jdbc.Sql;
+import org.springframework.test.context.ActiveProfiles;
 
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @DisplayNameGeneration(ReplaceUnderscores.class)
+@ActiveProfiles(value = "test")
 class ReservationTimeControllerTest {
+
+    @LocalServerPort
+    int port;
+
     private static void 예약_시간_생성(String startAt) {
         Map<String, String> params = new HashMap<>();
         params.put("startAt", startAt);
@@ -32,8 +38,13 @@ class ReservationTimeControllerTest {
                 .statusCode(201);
     }
 
+    @BeforeEach
+    void setUp() {
+        RestAssured.port = port;
+    }
+
     @Test
-    void 예약_시간이_정상적으로_생성(){
+    void 예약_시간이_정상적으로_생성() {
         예약_시간_생성("10:00");
 
         RestAssured.given().log().all()
@@ -80,11 +91,10 @@ class ReservationTimeControllerTest {
                 .then().log().all()
                 .statusCode(201);
 
-
         RestAssured.given().log().all()
                 .when().get("/reservations")
                 .then().log().all()
                 .statusCode(200)
-                .body("size()", is(4));
+                .body("size()", is(26));
     }
 }
