+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -6,17 +6,17 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
+import javax.sql.DataSource;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.web.server.ResponseStatusException;
-import roomescape.dao.ThemeDao;
-import roomescape.fake.FakeReservaionDao;
-import roomescape.fake.FakeReservationTimeDao;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseBuilder;
+import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseType;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
+import roomescape.dao.ThemeDao;
 import roomescape.dto.ReservationRequest;
 import roomescape.dto.ReservationResponse;
-import roomescape.fake.FakeThemeDao;
 import roomescape.model.ReservationTime;
 import roomescape.model.Theme;
 
@@ -26,11 +26,18 @@ class ReservationServiceTest {
     private ReservationTimeDao reservationTimeDao;
     private ThemeDao themeDao;
 
+    private JdbcTemplate jdbcTemplate;
+
     @BeforeEach
     void setUp() {
-        ReservationDao reservationDao = new FakeReservaionDao();
-        reservationTimeDao = new FakeReservationTimeDao();
-        themeDao = new FakeThemeDao();
+        DataSource dataSource = new EmbeddedDatabaseBuilder().setType(EmbeddedDatabaseType.H2)
+                .addScript("schema.sql")
+                .addScript("data.sql")
+                .build();
+        jdbcTemplate = new JdbcTemplate(dataSource);
+        ReservationDao reservationDao = new ReservationDao(jdbcTemplate);
+        reservationTimeDao = new ReservationTimeDao(jdbcTemplate);
+        themeDao = new ThemeDao(jdbcTemplate);
         reservationService = new ReservationService(reservationDao, reservationTimeDao, themeDao);
     }
 
@@ -38,38 +45,34 @@ class ReservationServiceTest {
     void 예약을_정상적으로_추가() {
         ReservationTime savedTime = reservationTimeDao.save(new ReservationTime(null, LocalTime.of(10, 0)));
         Theme savedTheme = themeDao.save(new Theme(null, "제목", "de", "th"));
-        ReservationRequest request = new ReservationRequest("이름", LocalDate.of(2025,12,16),savedTime.getId(), savedTheme.getId());
+        ReservationRequest request = new ReservationRequest("이름", LocalDate.of(2025, 12, 16), savedTime.getId(),
+                savedTheme.getId());
 
         ReservationResponse response = reservationService.addReservation(request);
 
         assertThat(response.id()).isNotNull();
         assertThat(response.name()).isEqualTo("이름");
-        assertThat(response.date()).isEqualTo(LocalDate.of(2025,12,16).toString());
+        assertThat(response.date()).isEqualTo(LocalDate.of(2025, 12, 16).toString());
     }
 
     @Test
     void 예약_리스트를_정상적으로_조회() {
-        ReservationTime savedTime = reservationTimeDao.save(new ReservationTime(null, LocalTime.of(10, 0)));
-        Theme savedTheme = themeDao.save(new Theme(null, "제목", "de", "th"));
-        ReservationRequest request = new ReservationRequest("이름", LocalDate.of(2025,12,16),savedTime.getId(),
-                savedTheme.getId());
-
-        reservationService.addReservation(request);
         List<ReservationResponse> reservations = reservationService.getReservations();
-        assertThat(reservations.size()).isEqualTo(1);
+        assertThat(reservations.size()).isEqualTo(25);
     }
 
     @Test
     void 예약을_정상적으로_삭제() {
         ReservationTime savedTime = reservationTimeDao.save(new ReservationTime(null, LocalTime.of(10, 0)));
         Theme savedTheme = themeDao.save(new Theme(null, "제목", "de", "th"));
-        ReservationRequest request = new ReservationRequest("이름", LocalDate.of(2025,12,16),savedTime.getId(), savedTheme.getId());
+        ReservationRequest request = new ReservationRequest("이름", LocalDate.of(2025, 12, 16), savedTime.getId(),
+                savedTheme.getId());
         ReservationResponse response = reservationService.addReservation(request);
 
         reservationService.deleteReservation(response.id());
         List<ReservationResponse> reservations = reservationService.getReservations();
 
-        assertThat(reservations).isEmpty();
+        assertThat(reservations).hasSize(25);
     }
 
     @Test
